+++ ios/cunning_document_scanner/Sources/cunning_document_scanner/CunningScannerOptions.swift
@@ -5,24 +5,33 @@ enum CunningScannerImageFormat: String {
     case png
 }
 
+enum CunningScannerSource: String {
+    case camera
+    case gallery
+    case cameraAndGallery = "camera_and_gallery"
+}
+
 struct CunningScannerOptions {
     let imageFormat: CunningScannerImageFormat
     let jpgCompressionQuality: Double
     let asPdf: Bool
     let isGalleryImportAllowed: Bool
+    let scannerSource: CunningScannerSource
 
     init() {
         self.imageFormat = .png
         self.jpgCompressionQuality = 1.0
         self.asPdf = false
         self.isGalleryImportAllowed = false
+        self.scannerSource = .camera
     }
 
-    init(imageFormat: CunningScannerImageFormat, jpgCompressionQuality: Double, asPdf: Bool, isGalleryImportAllowed: Bool) {
+    init(imageFormat: CunningScannerImageFormat, jpgCompressionQuality: Double, asPdf: Bool, isGalleryImportAllowed: Bool, scannerSource: CunningScannerSource) {
         self.imageFormat = imageFormat
         self.jpgCompressionQuality = jpgCompressionQuality
         self.asPdf = asPdf
         self.isGalleryImportAllowed = isGalleryImportAllowed
+        self.scannerSource = scannerSource
     }
 
     static func fromArguments(args: Any?) -> CunningScannerOptions {
@@ -32,6 +41,9 @@ struct CunningScannerOptions {
 
         let asPdf = (arguments["asPdf"] as? Bool) ?? false
         let isGalleryImportAllowed = (arguments["isGalleryImportAllowed"] as? Bool) ?? false
+        let sourceString = arguments["scannerSource"] as? String
+        let scannerSource = CunningScannerSource(rawValue: sourceString ?? "") ?? (isGalleryImportAllowed ? .cameraAndGallery : .camera)
+        
         let scannerOptionsDict = arguments["iosScannerOptions"] as? [String: Any]
         let imageFormat = (scannerOptionsDict?["imageFormat"] as? String) ?? "png"
         let jpgCompressionQuality = (scannerOptionsDict?["jpgCompressionQuality"] as? Double) ?? 1.0
@@ -40,7 +52,8 @@ struct CunningScannerOptions {
             imageFormat: CunningScannerImageFormat(rawValue: imageFormat) ?? .png,
             jpgCompressionQuality: jpgCompressionQuality,
             asPdf: asPdf,
-            isGalleryImportAllowed: isGalleryImportAllowed
+            isGalleryImportAllowed: isGalleryImportAllowed,
+            scannerSource: scannerSource
         )
     }
 }